Fix(gui/SslErrorDialog): implement parsing Strict-Transport-Security
This commit implements parsing the `Strict-Transport-Security` header
and the max-age parameter. It is then added to currentDateTime() and
saved with the account settings.
On future connections, this setting is loaded and if the server
presents an invalid certificate and the HSTS hint is not expired, the
connection fails without showing the dreaded "Cannot connect securely
to ..." dialog.

Signed-off-by: François Guerraz <francois@orographic.uk>