fix(msi): set the REBOOT property in the MSI
A value of `ReallySuppress` will not display a reboot prompt whenever
the `ForceReboot` action was run, or whenever a reboot was scheduled
(e.g. by `FilesInUse`).  Apparently this will not reboot the machine on
its own either.

`winget` sets this property in its manifest[0], and according to reports in
issue #5369 this seems to work.

See also: https://learn.microsoft.com/en-us/windows/win32/msi/reboot

[0]: https://github.com/microsoft/winget-pkgs/blob/0b40904d1480f67d48c6d8e2ac7f27b31aa566c7/manifests/n/Nextcloud/NextcloudDesktop/3.16.5/Nextcloud.NextcloudDesktop.installer.yaml#L16

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>