ci: change download url of appimagetool
The appimagetool as previously provided from the AppImageKit repo was
deprecated and removed.

The new versions as provided by the appimagetool repository also ship
with an updated runtime which from my understanding improve
compatibility (e.g. libfuse is now statically linked, so systems don't
need to have it present) and seem to slightly reduce the binary size as
well.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>