fix(syncstatussummary): do not attempt to fetch VFS state for accounts without VFS enabled
The socket server won't ever know anything about accounts that are not
set up with VFS, resulting in a "Some files could not be synced" warning
message.  There is no use in trying to fetch the state in that case.

Since the state of the File Provider is per-account, iterating over all
accounts is not necessary either.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>