fix: ignore directory listing entry for the iterated directory
The `ignoreFirst` bool had a short lifetime, by the time the lambda slot
handling the `LsColJob::directoryListingIterated` signal was called the
reference to that bool would report a completely different value.  This
could result in the first entry of a directory listing to not be ignored
properly.

This commit changes the way the slot checks whether it's currently
iterating over the very same directory that was just requested by
comparing the request path with the path from the response.

Fixes #8915

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>