fix(checksums): avoid crash on aborted checksum calculation job
In case a propagation job is aborted while it's about to start computing
the content checksum in a thread, the thread could still be started and
attempting to call `_checksumCalculator->calculate()`.  As the job was
aborted the ComputeChecksum instance has already been deleted along with
its ChecksumCalculator instance, and the thread ends up calling
`calculate()` on an invalid pointer.

I managed to occasionally reproduce this through:

    auto computeChecksum = new ComputeChecksum();
    computeChecksum->setChecksumType("MD5");
    computeChecksum->start("/dev/zero");
    delete computeChecksum;

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>