fix: Metadata lookup for items with hashtag in name.
This surfaced in-house for myself when I tried to navigate to a subfolder which is named something like "Ticket#123" and the folder content enumeration failed.
It turned out that the URL operations for deriving the parent remote URL were not taking the URL fragment into account.
The strange thing is: I could have sworn that this worked before. Maybe the underlying frameworks changed in behavior of their APIs.

Signed-off-by: Iva Horn <iva.horn@nextcloud.com>