WIP: fix(win32): try to set ACLs without propagating them
apparently `SetFileSecurity` did not do that, unlike the "new" (since
Windows 2000) replacement `SetSecurityInfo`/`SetNamedSecurityInfo`.

`SetSecurityInfo` should not perform the propagation if the handle is
opened with an access mask value of `MAXIMUM_ALLOWED`, so let's give
that a try.

also removed the restriction on the FILE_WRITE_ATTRIBUTES permission so
that `FileSystem::setFileReadOnly` has an easier time modifying the
basic attribute

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>